Bug 1459785 - Update ANGLE to chromium/3396.
MozReview-Commit-ID: EA39lUfXuPI

UltraBlame original commit: b7c91a6f1b0a72da63f430d8f6c57d5374d7e0a7

+++ gfx/angle/checkout/src/libANGLE/renderer/d3d/BufferD3D.cpp
@@ -160,10 +160,11 @@ void BufferD3D::invalidateStaticData(const gl::Context *context)
 }
 
 
-void BufferD3D::promoteStaticUsage(const gl::Context *context, int dataSize)
+void BufferD3D::promoteStaticUsage(const gl::Context *context, size_t dataSize)
 {
     if (mUsage == D3DBufferUsage::DYNAMIC)
     {
+        
         mUnmodifiedDataUse += dataSize;
 
         if (mUnmodifiedDataUse > 3 * getSize())
